更新
1、公司可对活动进行编辑和删除
