+++ app/src/main/java/com/example/dell/c1ean/Fragment/Company/CompanyHomePageFragment.java
@@ -1,5 +1,6 @@
 package com.example.dell.c1ean.Fragment.Company;
 
+import android.content.Intent;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
@@ -48,5 +49,4 @@ public class CompanyHomePageFragment extends Fragment{
 
         return view;
     }
-
 }
