+++ app/src/main/res/layout/company_events.xml
@@ -4,18 +4,25 @@
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
     xmlns:app="http://schemas.android.com/apk/res-auto">
-    <include layout="@layout/title" android:id="@+id/title"/>
-<TextView
-    android:layout_below="@id/title"
-    android:id="@+id/name"
-    android:background="@color/colorWhite"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:padding="@dimen/buttonPadding"
-    android:text="我发布过的活动"
-    android:gravity="center_vertical|center_horizontal"/>
+
+    <LinearLayout
+        android:id="@+id/title"
+        android:orientation="vertical"
+        android:background="@color/colorWhite"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content">
+        <TextView
+            android:id="@+id/title_bar"
+            android:layout_width="match_parent"
+            android:gravity="center_horizontal|center_vertical"
+            android:layout_height="wrap_content"
+            android:text="我发布的活动"
+            android:layout_marginBottom="@dimen/marginTop"
+            android:textColor="@color/colorTheme"/>
+        <include layout="@layout/line"/>
+    </LinearLayout>
     <ScrollView
-        android:layout_below="@id/name"
+        android:layout_below="@id/title"
         android:id="@+id/scrollView"
         android:layout_width="match_parent"
         android:layout_height="match_parent">
