+++ app/src/main/res/layout/company_homepage.xml
@@ -3,7 +3,23 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true">
-    <include layout="@layout/title" android:id="@+id/title"/>
+
+    <LinearLayout
+        android:id="@+id/title"
+        android:orientation="vertical"
+        android:background="@color/colorWhite"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content">
+        <TextView
+            android:id="@+id/title_bar"
+            android:layout_width="match_parent"
+            android:gravity="center_horizontal|center_vertical"
+            android:layout_height="wrap_content"
+            android:text="首页"
+            android:layout_marginBottom="@dimen/marginTop"
+            android:textColor="@color/colorTheme"/>
+        <include layout="@layout/line"/>
+    </LinearLayout>
     <ScrollView
         android:layout_below="@+id/title"
         android:layout_width="match_parent"
