+++ app/src/main/res/layout/mainpage.xml
@@ -2,8 +2,7 @@
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:fitsSystemWindows="true"
-    android:gravity="center_vertical|center_horizontal">
+    android:fitsSystemWindows="true">
     <com.ycl.tabview.library.TabView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
