+++ app/src/main/res/layout/notice_page.xml
@@ -20,9 +20,9 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:text="用户"
-        android:textColor="@color/colorOrange"/>
+        android:textColor="@color/colorWhite"/>
     <TextView
-        android:textColor="@color/colorOrange"
+        android:textColor="@color/colorWhite"
         android:id="@+id/user_tel"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
@@ -30,18 +30,18 @@
     <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:textColor="@color/colorOrange"
+        android:textColor="@color/colorWhite"
         android:text="预定了"/>
     <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:textColor="@color/colorOrange"
+        android:textColor="@color/colorWhite"
         android:text="家具清洗"
         android:id="@+id/activity_type"/>
     <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:text="服务"
-        android:textColor="@color/colorOrange"/>
+        android:textColor="@color/colorWhite"/>
     </LinearLayout>
 </LinearLayout>
