+++ app/src/main/res/layout/user_homepage.xml
@@ -2,6 +2,7 @@
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
+    android:background="@color/colorTheme"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true">
 
@@ -14,10 +15,10 @@
         <TextView
             android:id="@+id/title_bar"
             android:layout_width="match_parent"
-            android:padding="10dp"
             android:gravity="center_horizontal|center_vertical"
             android:layout_height="wrap_content"
             android:text="首页"
+            android:layout_marginBottom="@dimen/marginTop"
             android:textColor="@color/colorTheme"/>
         <include layout="@layout/line"/>
     </LinearLayout>
@@ -40,7 +41,6 @@
         app:rippleColor="@color/colorGray" />
 
     <ScrollView
-        android:padding="@dimen/marginTop"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:layout_below="@id/title"
@@ -48,6 +48,7 @@
         android:scrollbars="none">
 
         <RelativeLayout
+            android:padding="@dimen/marginTop"
             android:layout_width="match_parent"
             android:layout_height="wrap_content">
 
@@ -74,7 +75,6 @@
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:layout_below="@id/line"
-                android:layout_marginTop="10dp"
                 app:vsl_animDuration="800"
                 app:vsl_isCusDuration="true"
                 app:vsl_sleepTime="3000" />
@@ -84,8 +84,8 @@
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:layout_below="@id/buys_scroll"
-                android:layout_marginTop="10dp"
-                android:background="@color/colorWhite"
+                android:padding="@dimen/marginTop"
+                android:background="@drawable/modle_shape"
                 android:orientation="vertical">
 
                 <TextView
