+++ app/src/main/res/layout/user_order_page.xml
@@ -6,42 +6,100 @@
 
     <LinearLayout
         android:id="@+id/title"
-        android:orientation="vertical"
-        android:background="@color/colorWhite"
         android:layout_width="match_parent"
-        android:layout_height="wrap_content">
+        android:layout_height="wrap_content"
+        android:background="@color/colorWhite"
+        android:orientation="vertical">
+
         <TextView
             android:id="@+id/title_bar"
             android:layout_width="match_parent"
-            android:padding="10dp"
-            android:gravity="center_horizontal|center_vertical"
             android:layout_height="wrap_content"
+            android:gravity="center_horizontal|center_vertical"
+            android:layout_marginBottom="@dimen/marginTop"
             android:text="我的订单"
-            android:textColor="@color/colorTheme"/>
-        <include layout="@layout/line"/>
+            android:textColor="@color/colorTheme" />
+
+        <include layout="@layout/line" />
     </LinearLayout>
 
-    <TabHost
-        android:id="@android:id/tabhost"
+    <!--<TabHost-->
+    <!--android:id="@android:id/tabhost"-->
+    <!--android:layout_width="match_parent"-->
+    <!--android:layout_height="match_parent"-->
+    <!--android:layout_below="@id/title">-->
+
+    <!--<LinearLayout-->
+    <!--android:layout_width="match_parent"-->
+    <!--android:layout_height="match_parent"-->
+    <!--android:orientation="vertical">-->
+
+    <!--<TabWidget-->
+    <!--android:background="@color/colorWhite"-->
+    <!--android:id="@android:id/tabs"-->
+    <!--android:layout_width="match_parent"-->
+    <!--android:layout_height="wrap_content" />-->
+
+    <!--<FrameLayout-->
+    <!--android:id="@android:id/tabcontent"-->
+    <!--android:layout_width="match_parent"-->
+    <!--android:layout_height="match_parent" />-->
+    <!--</LinearLayout>-->
+    <!--</TabHost>-->
+    <LinearLayout
+        android:id="@+id/line"
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_below="@id/title">
+        android:layout_height="wrap_content"
+        android:layout_below="@id/title"
+        android:orientation="horizontal">
 
-        <LinearLayout
+        <TextView
+            android:id="@+id/all_order"
             android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_weight="1"
+            android:background="@color/colorWhite"
+            android:padding="20dp"
+            android:text="所有订单"
+            android:textAlignment="center" />
+
+        <View
+            android:layout_width="1dp"
             android:layout_height="match_parent"
-            android:orientation="vertical">
+            android:layout_marginBottom="5dp"
+            android:layout_marginTop="5dp"
+            android:background="@color/colorGray" />
 
-            <TabWidget
-                android:background="@color/colorWhite"
-                android:id="@android:id/tabs"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content" />
+        <TextView
+            android:id="@+id/wait_pay"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_weight="1"
+            android:background="@color/colorWhite"
+            android:padding="20dp"
+            android:text="待支付订单"
+            android:textAlignment="center" />
 
-            <FrameLayout
-                android:id="@android:id/tabcontent"
-                android:layout_width="match_parent"
-                android:layout_height="match_parent" />
-        </LinearLayout>
-    </TabHost>
+        <View
+            android:layout_width="1dp"
+            android:layout_height="match_parent"
+            android:layout_marginBottom="5dp"
+            android:layout_marginTop="5dp"
+            android:background="@color/colorGray" />
+
+        <TextView
+            android:id="@+id/wait_evaluate"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_weight="1"
+            android:background="@color/colorWhite"
+            android:padding="20dp"
+            android:text="待评价订单"
+            android:textAlignment="center" />
+    </LinearLayout>
+    <FrameLayout
+        android:id="@+id/order_page"
+        android:layout_below="@id/line"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"></FrameLayout>
 </RelativeLayout>
