+++ app/src/main/res/layout/user_page.xml
@@ -14,7 +14,7 @@
         <TextView
             android:id="@+id/title_bar"
             android:layout_width="match_parent"
-            android:padding="10dp"
+            android:layout_marginBottom="@dimen/marginTop"
             android:gravity="center_horizontal|center_vertical"
             android:layout_height="wrap_content"
             android:text="个人中心"
