+++ app/src/main/res/layout/user_wait_pay_orders.xml
@@ -13,7 +13,7 @@
         android:id="@+id/valuable"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="您还没有订单哦"
+        android:text="您没有待付款订单哦"
         android:layout_centerInParent="true"
         android:visibility="visible"/>
 </RelativeLayout>
